login + user class abstraction
