+++ app/__init__.py
@@ -4,17 +4,27 @@ from flask_login import LoginManager, UserMixin, login_user, login_required, log
 # from werkzeug.security import generate_password_hash, check_password_hash
 from bcrypt import hashpw, checkpw, gensalt
 
-def create_app(name):
-    app = Flask(name, template_folder='app/templates')
-    app.static_folder = 'app/static'
-    app.config['SECRET_KEY'] = 'TREND IS THE BEST'
+db = SQLAlchemy()
+login_manager = LoginManager()
 
-    # app = Flask(__name__)
-    # app.config['SECRET_KEY'] = 'your_secret_key'
+def create_app():
+    app = Flask(__name__)
+    app.config['SECRET_KEY'] = 'TREND IS THE BEST'
     app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///users.db'
-    db = SQLAlchemy(app)
-    login_manager = LoginManager(app)
-    login_manager.login_view = 'login'
+
+    db.init_app(app)
+    login_manager.init_app(app)
+    login_manager.login_view = 'login.login'
+    # app = Flask(name, template_folder='app/templates')
+    # app.static_folder = 'app/static'
+    # app.config['SECRET_KEY'] = 'TREND IS THE BEST'
+
+    # # app = Flask(__name__)
+    # # app.config['SECRET_KEY'] = 'your_secret_key'
+    # app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///users.db'
+    # db = SQLAlchemy(app)
+    # login_manager = LoginManager(app)
+    # login_manager.login_view = 'login'
 
     @login_manager.user_loader
     def load_user(user_id):
@@ -32,19 +42,24 @@ def create_app(name):
     def home():
         return render_template("home.html")
     
-    @app.route('/login', methods=['GET', 'POST'])
-    def login():
-        if request.method == 'POST':
-            username = request.form['username']
-            password = request.form['password'].encode('utf-8')
-            user = User.query.filter_by(username=username).first()
+    from app.src.auth.login import login_bp
+    from app.src.auth.logout import logout_bp
+    app.register_blueprint(login_bp, url_prefix='/auth')
+    app.register_blueprint(logout_bp, url_prefix='/auth')
 
-            if user and checkpw(password, user.password.encode('utf-8')):
-                login_user(user)
-                return redirect(url_for('home'))
-            else:
-                flash('Invalid username or password.')
-        return render_template('login.html')
+    # @app.route('/login', methods=['GET', 'POST'])
+    # def login():
+    #     if request.method == 'POST':
+    #         username = request.form['username']
+    #         password = request.form['password'].encode('utf-8')
+    #         user = User.query.filter_by(username=username).first()
+
+    #         if user and checkpw(password, user.password.encode('utf-8')):
+    #             login_user(user)
+    #             return redirect(url_for('home'))
+    #         else:
+    #             flash('Invalid username or password.')
+    #     return render_template('login.html')
     
     @app.route('/register', methods=['GET', 'POST'])
     def register():
@@ -58,9 +73,6 @@ def create_app(name):
             flash('Registration successful! You can now log in.')
             return redirect(url_for('login'))
         return render_template('register.html')
-
-    from app.src.auth.logout import auth_bp as logout_bp
-    app.register_blueprint(logout_bp, url_prefix='/auth')
     
     # @app.route('/logout')
     # @login_required
