+++ app/src/db/user.py
@@ -1,0 +1,11 @@
+from flask_sqlalchemy import SQLAlchemy
+from flask_login import UserMixin
+from app import db
+
+class User(db.Model, UserMixin):
+    __tablename__ = 'user'
+    __table_args__ = {'extend_existing': True}
+    
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(150), nullable=False, unique=True)
+    password = db.Column(db.String(150), nullable=False)
