+++ run.py
@@ -2,5 +2,5 @@ from app import create_app
 from flask import Flask, render_template
 
 if __name__ == '__main__':
-    app = create_app(__name__)
+    app = create_app()
     app.run(debug=True)
